backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.36
alembic==1.12.1

# Data Validation
pydantic==2.8.2
pydantic-settings==2.3.2

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# CORS Support
python-multipart==0.0.6

# Date/Time Handling
python-dateutil==2.8.2

# Optional but Recommended
# Database Driver (if you upgrade to PostgreSQL later)
# psycopg2-binary==2.9.9

# Development/Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1